refactor(mongodb): update MongoDB provider to use a singleton pattern and improve connection handling
refactor(routes): rename favorite-schedules and schedules routes for better organization and update to use new MongoDB provider
