+++ server/provider/mongodb.provider.ts
@@ -1,22 +1,17 @@
 import { MongoClient, Db } from 'mongodb'
 import { useRuntimeConfig } from '#imports'
+import type { H3Event, EventHandlerRequest } from 'h3'
 
-let cachedDb: Db | null = null
+let dbInstance: Db | null = null
 
-export async function useMongoDB() {
-  if (cachedDb) {
-    return { db: cachedDb }
+export async function useMongoDB(event: H3Event<EventHandlerRequest>): Promise<Db> {
+  if (dbInstance) {
+    return dbInstance
   }
-
-  const config = useRuntimeConfig()
-  const client = new MongoClient(config.mongodb.uri, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  })
-
-  await client.connect()
-  const db = client.db(config.mongodb.dbName)
-  cachedDb = db
-
-  return { db }
+  const config = useRuntimeConfig(event)
+  const client = new MongoClient(config.mongodb.uri)
+  const clientConnected = await client.connect()
+  const db = clientConnected.db(config.mongodb.dbName)
+  dbInstance = db
+  return db
 }
