+++ server/routes/favorite-schedules/index.post.ts
@@ -1,13 +1,12 @@
 import { z } from 'zod'
-import { favoriteScheduleSchema } from '../schemas/favorite-schedule.schema'
-import { getMongoClient } from '../provider/mongodb.provider'
+import { favoriteScheduleSchema } from '../../schemas/favorite-schedule.schema'
+import { useMongoDB } from '../../provider/mongodb.provider'
 
 export default defineEventHandler(async (event) => {
   const body = await readBody(event)
   const parsedBody = favoriteScheduleSchema.parse(body)
 
-  const client = await getMongoClient()
-  const db = client.db()
+  const db = await useMongoDB(event)
   const collection = db.collection('favoriteSchedules')
 
   await collection.insertOne(parsedBody)
