+++ server/routes/schedules/index.get.ts
@@ -1,8 +1,7 @@
-import { ObjectId } from 'mongodb'
-import { useMongoDB } from '../provider/mongodb.provider'
+import { useMongoDB } from '../../provider/mongodb.provider'
 
 export default defineEventHandler(async (event) => {
-  const { db } = await useMongoDB()
+  const db = await useMongoDB(event)
   const collection = db.collection('schedules')
 
   const schedules = await collection.find().toArray()
