+++ server/routes/schedules/index.post.ts
@@ -1,13 +1,11 @@
-import { z } from 'zod'
-import { ObjectId } from 'mongodb'
-import { useMongoDB } from '../provider/mongodb.provider'
-import { scheduleSchema } from '../schemas/schedule.schema'
+import { readBody } from 'h3'
+import { useMongoDB } from '../../provider/mongodb.provider'
+import { scheduleSchema } from '../../schemas/schedule.schema'
 
 export default defineEventHandler(async (event) => {
   const body = await readBody(event)
   const parsedBody = scheduleSchema.parse(body)
-
-  const { db } = await useMongoDB()
+  const db = await useMongoDB(event)
   const collection = db.collection('schedules')
 
   const result = await collection.insertOne({
