+++ server/routes/subjects/index.get.ts
@@ -1,8 +1,8 @@
 import { ObjectId } from 'mongodb'
-import { useMongoDB } from '../provider/mongodb.provider'
+import { useMongoDB } from '../../provider/mongodb.provider'
 
 export default defineEventHandler(async (event) => {
-  const { db } = await useMongoDB()
+  const db = await useMongoDB(event)
   const collection = db.collection('subjects')
 
   const subjects = await collection.find().toArray()
