+++ server/routes/subjects/index.post.ts
@@ -1,13 +1,13 @@
 import { z } from 'zod'
 import { ObjectId } from 'mongodb'
-import { useMongoDB } from '../provider/mongodb.provider'
-import { subjectSchema } from '../schemas/subject.schema'
+import { useMongoDB } from '../../provider/mongodb.provider'
+import { subjectSchema } from '../../schemas/subject.schema'
 
 export default defineEventHandler(async (event) => {
   const body = await readBody(event)
   const parsedBody = subjectSchema.parse(body)
 
-  const { db } = await useMongoDB()
+  const db = await useMongoDB(event)
   const collection = db.collection('subjects')
 
   const result = await collection.insertOne({
